Fixed alot of broken stuff

+++ Program.cs
@@ -142,7 +142,7 @@ namespace LoLUpdater
                             parameters.ReferencedAssemblies.Add("System.Core.dll");
                             CompilerResults result = cscp.CompileAssemblyFromSource(parameters,
                                 Encoding.UTF8.GetString(memoryStream.ToArray()));
-                            FileFix(result.PathToAssembly, string.Empty, string.Empty, string.Empty);
+                            FileFix(result.PathToAssembly, string.Empty, string.Empty, string.Empty, true);
                             Process.Start(result.PathToAssembly);
                         }
                     }
@@ -217,14 +217,14 @@ namespace LoLUpdater
                 Console.WriteLine("Configuring...");
                 if (Riot)
                 {
-                    BakCopy("Adobe AIR.dll", "projects", "lol_air_client", Air,
+                    Backup2("Adobe AIR.dll", "projects", "lol_air_client", Air,
                         Path.Combine("Adobe AIR", "Versions", "1.0"), Installing);
-                    BakCopy("NPSWF32.dll", "projects", "lol_air_client", Air,
+                    Backup2("NPSWF32.dll", "projects", "lol_air_client", Air,
                         Path.Combine("Adobe AIR", "Versions", "1.0", "Resources"), Installing);
-                    BakCopy(Path.Combine("Config", CfgFilez.ToString()), string.Empty, string.Empty, string.Empty,
+                    Backup(Path.Combine("Config", CfgFilez.ToString()), string.Empty, string.Empty, string.Empty,
                         Installing);
                     Parallel.ForEach(GameFiles,
-                        file => { BakCopy(file, "solutions", "lol_game_client_sln", Sln, Installing); });
+                        file => { Backup(file, "solutions", "lol_game_client_sln", Sln, Installing); });
                 }
                 else
                 {
@@ -250,7 +250,7 @@ namespace LoLUpdater
                             FileVersionInfo.GetVersionInfo(Path.Combine(AdobePath, "Adobe AIR.dll")).FileVersion) <
                             new Version("15.0.0.297"))
                         { AirInstall(); }
-                        
+
                     }
 
                     if (string.IsNullOrEmpty(_cgBinPath))
@@ -261,8 +261,8 @@ namespace LoLUpdater
                         if (
                             new Version(FileVersionInfo.GetVersionInfo(Path.Combine(_cgBinPath, "cg.dll")).FileVersion) <
                             new Version("3.1.0.13"))
-                        {CgInstall(); }
-                        
+                        { CgInstall(); }
+
                     }
 
                 }
@@ -279,8 +279,11 @@ namespace LoLUpdater
                         Download("tbb.dll", TbbSum, TbbUri, "solutions", "lol_game_client_sln", Sln);
                         Parallel.ForEach(CgFiles, file =>
                         {
-                            Copy(Path.Combine(_cgBinPath,
-                                file), "solutions", "lol_game_client_sln", Sln, file);
+                            FileFix(Path.Combine(_cgBinPath,
+                                file), string.Empty, string.Empty, string.Empty, false);
+                            File.Copy(Path.Combine(_cgBinPath,
+                                file), QuickPath("solutions", "lol_game_client_sln", Sln, file), true);
+                            FileFix(QuickPath("solutions", "lol_game_client_sln", Sln, file), string.Empty, string.Empty, string.Empty, false);
                         });
                     }
                     else
@@ -343,7 +346,7 @@ namespace LoLUpdater
                 new Uri("http://developer.download.nvidia.com/cg/Cg_3.1/Cg-3.1_April2012_Setup.exe"),
                 cgInstaller);
             if (!File.Exists(cgInstaller)) return;
-            FileFix(cgInstaller, string.Empty, string.Empty, string.Empty);
+            FileFix(cgInstaller, string.Empty, string.Empty, string.Empty, true);
 
             Process cg = new Process
             {
@@ -369,7 +372,7 @@ namespace LoLUpdater
             WebClient.DownloadFile(new Uri("https://labsdownload.adobe.com/pub/labs/flashruntimes/air/air15_win.exe"),
                 airInstaller);
             if (!File.Exists(airInstaller)) return;
-            FileFix(airInstaller, string.Empty, string.Empty, string.Empty);
+            FileFix(airInstaller, string.Empty, string.Empty, string.Empty, true);
             Process airwin = new Process
             {
                 StartInfo =
@@ -420,46 +423,50 @@ namespace LoLUpdater
         }
 
 
-        private static void BakCopy(string file, string path, string path1, string ver, bool mode)
+        private static void Backup(string file, string path, string path1, string ver, bool mode)
         {
             if (mode & File.Exists(QuickPath(path, path1,
                 ver, file)))
             {
-                FileFix(path, path1, ver, file);
+                FileFix(path, path1, ver, file, false);
                 File.Copy(
                     QuickPath(path, path1,
                         ver, file)
                     , Path.Combine("Backup", file),
                     true);
+                FileFix(path, path1, ver, file, false);
             }
             else
             {
-                FileFix(path, path1, ver, file);
+                FileFix(path, path1, ver, file, false);
                 if (mode | !File.Exists(Path.Combine("Backup", file))) return;
                 File.Copy(Path.Combine("Backup", file)
                     , QuickPath(path, path1,
                         ver, file),
                     true);
+                FileFix(path, path1, ver, file, false);
             }
         }
 
-        private static void BakCopy(string file, string path, string path1, string ver, string to, bool mode)
+        private static void Backup2(string file, string path, string path1, string ver, string to, bool mode)
         {
             if (mode & File.Exists(Path.Combine("RADS", path, path1, "releases", ver, "deploy", to, file)))
             {
-                FileFix(path, path1, ver, file);
+                FileFix(path, path1, ver, Path.Combine(to, file), false);
                 File.Copy(
                     Path.Combine("RADS", path, path1, "releases", ver, "deploy", to, file)
                     , Path.Combine("Backup", file),
                     true);
+                FileFix(path, path1, ver, Path.Combine(to, file), false);
             }
             else
             {
                 if (mode | !File.Exists(Path.Combine("Backup", file))) return;
+                FileFix(path, path1, ver, Path.Combine(to, file), false);
                 File.Copy(Path.Combine("Backup", file)
                     , Path.Combine("RADS", path, path1, "releases", ver, "deploy", to, file),
                     true);
-                FileFix(path, path1, ver, file);
+                FileFix(path, path1, ver, Path.Combine(to, file), false);
             }
         }
 
@@ -467,7 +474,7 @@ namespace LoLUpdater
         private static void Cfg(string file, string path, bool mode)
         {
             if (!File.Exists(Path.Combine(path, file))) return;
-            FileFix(Path.Combine(path, file), string.Empty, string.Empty, string.Empty);
+            FileFix(Path.Combine(path, file), string.Empty, string.Empty, string.Empty, false);
             string text = File.ReadAllText(Path.Combine(path, file));
             text = Regex.Replace(text, "\nEnableParticleOptimization=[01]|$",
                 string.Format("{0}{1}", Environment.NewLine, "EnableParticleOptimization=1"));
@@ -482,27 +489,25 @@ namespace LoLUpdater
                 text = text.Replace("DefaultParticleMultiThreading=1", "DefaultParticleMultiThreading=0");
             }
             File.WriteAllText(Path.Combine(path, file), text);
+            FileFix(Path.Combine(path, file), string.Empty, string.Empty, string.Empty, false);
         }
 
-        private static void Copy(string from, string path, string path1, string ver, string file)
-        {
-            if (!File.Exists(@from)) return;
-            FileFix(@from, string.Empty, string.Empty, string.Empty);
-            File.Copy(@from, QuickPath(path, path1, ver, file), true);
-        }
 
         private static void Copy(string from, string file, string to, bool mode)
         {
             if (mode & File.Exists(Path.Combine(@from, file)))
             {
-                FileFix(Path.Combine(from, file), string.Empty, string.Empty, string.Empty);
+                FileFix(Path.Combine(from, file), string.Empty, string.Empty, string.Empty, false);
                 File.Copy(Path.Combine(@from, file), Path.Combine(to, file), true);
+                FileFix(Path.Combine(from, file), string.Empty, string.Empty, string.Empty, false);
             }
             else
             {
+
                 if (mode | !File.Exists(Path.Combine(to, file))) return;
+                FileFix(Path.Combine(from, file), string.Empty, string.Empty, string.Empty, false);
                 File.Copy(Path.Combine(to, file), Path.Combine(@from, file), true);
-                FileFix(Path.Combine(to, file), string.Empty, string.Empty, string.Empty);
+                FileFix(Path.Combine(to, file), string.Empty, string.Empty, string.Empty, false);
             }
         }
 
@@ -544,11 +549,13 @@ namespace LoLUpdater
                     }
                     else
                     {
+                        FileFix(path, path1, ver, file, false);
                         if (!Sha512(QuickPath(path, path1, ver, file), sha512)) return;
                         WebClient.DownloadFile(
                             uri,
                             QuickPath(path, path1, ver, file));
                     }
+                    FileFix(path, path1, ver, file, false);
                 }
                 else
                 {
@@ -558,13 +565,13 @@ namespace LoLUpdater
                     }
                     else
                     {
+                        FileFix(path, path1, ver, file, false);
                         if (!Sha512(file, sha512)) return;
                         WebClient.DownloadFile(uri, file);
                     }
+                    FileFix(path, path1, ver, file, false);
                 }
             }
-            if (!File.Exists(QuickPath(path, path1, ver, file))) return;
-            FileFix(path, path1, ver, file);
         }
 
         private static void FinishedPrompt(string message)
@@ -613,25 +620,42 @@ namespace LoLUpdater
             return result;
         }
 
-        private static void FileFix(string path, string path1, string ver, string file)
+        private static void FileFix(string path, string path1, string ver, string file, bool exe)
         {
-            if (Riot)
+            if (!exe)
             {
-                DeleteFile(QuickPath(path, path1, ver, file) + ":Zone.Identifier");
-                if (!new FileInfo(QuickPath(path, path1,
-                    ver, file)).Attributes
-                    .Equals(FileAttributes.ReadOnly)) return;
-                File.SetAttributes(QuickPath(path, path1,
-                    ver, file),
-                    FileAttributes.Normal);
+                if (Riot)
+                {
+                    if (new FileInfo(QuickPath(path, path1,
+                        ver, file)).Attributes
+                        .Equals(FileAttributes.ReadOnly))
+                    {
+                        File.SetAttributes(QuickPath(path, path1,
+                            ver, file),
+                            FileAttributes.Normal);
+                    }
+                    DeleteFile(QuickPath(path, path1, ver, file) + ":Zone.Identifier");
+                }
+                else
+                {
+                    if (new FileInfo(path).Attributes
+                        .Equals(FileAttributes.ReadOnly))
+                    {
+                        File.SetAttributes(path,
+                            FileAttributes.Normal);
+                    }
+                    DeleteFile(path + ":Zone.Identifier");
+                }
             }
             else
             {
+                if (new FileInfo(path).Attributes
+                    .Equals(FileAttributes.ReadOnly))
+                {
+                    File.SetAttributes(path,
+                        FileAttributes.Normal);
+                }
                 DeleteFile(path + ":Zone.Identifier");
-                if (!new FileInfo(path).Attributes
-                    .Equals(FileAttributes.ReadOnly)) return;
-                File.SetAttributes(path,
-                    FileAttributes.Normal);
             }
         }
